Rename public controller
